+++ gammapy/stats/__init__.py
@@ -1,6 +1,6 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Statistics."""
-from .significance import *
+from .normal import *
 from .data import *
 from .feldman_cousins import *
 from .fit_statistics import *
