+++ gammapy/stats/normal.py
